# Web/wwwroot/docs/api-specification.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "GrowRoomEnvironment WebAPI",
    "description": "GrowRoomEnvironment WebAPI Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/EventCondition": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_Post",
        "requestBody": {
          "x-name": "eventConditionVM",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConditionViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{pageNumber}/{pageSize}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetAllPaged",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{eventConditionId}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_Get",
        "parameters": [
          {
            "name": "eventConditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_Delete",
        "parameters": [
          {
            "name": "eventConditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_Put",
        "parameters": [
          {
            "name": "eventConditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "eventConditionVM",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConditionViewModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_Patch",
        "parameters": [
          {
            "name": "eventConditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patch",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{eventId}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetByEventId",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{eventId}/{pageNumber}/{pageSize}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetByEventIdPaged",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{dataPointId}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetByDataPointId",
        "parameters": [
          {
            "name": "dataPointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventCondition/{dataPointId}/{pageNumber}/{pageSize}": {
      "get": {
        "tags": [
          "EventCondition"
        ],
        "operationId": "EventCondition_GetByDataPointIdPaged",
        "parameters": [
          {
            "name": "dataPointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventConditionViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/{getDisabled}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_GetAll",
        "parameters": [
          {
            "name": "getDisabled",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/{pageNumber}/{pageSize}/{getDisabled}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_GetAllPaged",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "getDisabled",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/{actionDeviceId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_GetByActionDeviceId",
        "parameters": [
          {
            "name": "actionDeviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/{actionDeviceId}/{pageNumber}/{pageSize}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_GetByActionDeviceIdPaged",
        "parameters": [
          {
            "name": "actionDeviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventViewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/{eventId}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_Get",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_Delete",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_Put",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "eventVM",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventViewModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Event"
        ],
        "operationId": "Event_Patch",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patch",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventViewModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "Operation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OperationBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "value": {
                "nullable": true
              }
            }
          }
        ]
      },
...
      "EventViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationViewModelBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "eventId": {
                "type": "integer",
                "format": "int32"
              },
              "name": {
                "type": "string",
                "nullable": true
              },
              "actionDeviceId": {
                "type": "integer",
                "format": "int32"
              },
              "actionDevice": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ActionDeviceViewModel"
                  }
                ]
              },
              "state": {
                "$ref": "#/components/schemas/ActionDeviceStates"
              },
              "stateDescription": {
                "type": "string",
                "nullable": true
              },
              "isEnabled": {
                "type": "boolean"
              },
              "eventConditions": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/EventConditionViewModel"
                }
              }
            }
          }
        ]
      },
...
      "EventConditionViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationViewModelBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "eventConditionId": {
                "type": "integer",
                "format": "int32"
              },
              "eventId": {
                "type": "integer",
                "format": "int32"
              },
              "event": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EventViewModel"
                  }
                ]
              },
              "dataPointId": {
                "type": "integer",
                "format": "int32"
              },
              "dataPoint": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DataPointViewModel"
                  }
                ]
              },
              "operator": {
                "$ref": "#/components/schemas/Operators"
              },
              "operatorDescription": {
                "type": "string",
                "nullable": true
              },
              "value": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
...
    }
  }
}